+++ lib/global/actionListener.dart
@@ -46,30 +46,99 @@ class GlobalLeftScroll {
     }
   }
 
-  /// 删除目标行
-  Future<void> removeRowWithAnimation(
+  /// 删除目标行，带有动画
+  Future<void> removeRowAsync(
     LeftScrollCloseTag? tag,
     Key? key, {
-    Function? onRemove,
+    Future<bool?> Function()? onRemove,
   }) async {
     if (tag == null) return;
     if (map[tag] == null) return;
     // targetStatus(tag, key)?.value = LeftScrollStatus.close;
-    targetStatus(tag, key)?.value = LeftScrollStatus.remove;
+    targetStatus(tag, key)?.value = LeftScrollStatus.removing;
     await Future.delayed(Duration(milliseconds: 300));
-    onRemove?.call();
+    final removeResult = onRemove?.call();
+    if (removeResult == true) {
+      targetStatus(tag, key)?.value = LeftScrollStatus.removed;
+    } else {
+      targetStatus(tag, key)?.value = LeftScrollStatus.close;
+    }
+  }
+
+  /// 删除目标行，带有动画
+  Future<void> removeRowWithAnimation(
+    LeftScrollCloseTag? tag,
+    Key? key, {
+    required Function onRemove,
+  }) async {
+    if (tag == null) return;
+    if (map[tag] == null) return;
+    targetStatus(tag, key)?.value = LeftScrollStatus.removing;
+    await Future.delayed(Duration(milliseconds: 300));
+    onRemove.call();
     targetStatus(tag, key)?.value = LeftScrollStatus.removed;
   }
+}
+
+class LeftScrollController {
+  final LeftScrollCloseTag tag;
+  final Key key;
+
+  LeftScrollController({required this.tag, required this.key});
+
+  LeftScrollStatusCtrl? get status =>
+      GlobalLeftScroll.instance.targetStatus(tag, key);
+
+  to(LeftScrollStatus newStatus) {
+    status?.value = newStatus;
+  }
+
+  /// 删除目标行，带有动画
+  Future<void> open(
+    Future<bool?> Function()? onRemove,
+  ) async {
+    final targetCtrl = GlobalLeftScroll.instance.targetStatus(tag, key);
+    targetCtrl?.value = LeftScrollStatus.open;
+  }
+
+  /// 删除目标行，带有动画
+  Future<void> close(
+    Future<bool?> Function()? onRemove,
+  ) async {
+    final targetCtrl = GlobalLeftScroll.instance.targetStatus(tag, key);
+    targetCtrl?.value = LeftScrollStatus.close;
+  }
+
+  /// 删除目标行，带有动画
+  Future<void> remove(
+    Future<bool?> Function()? onRemove,
+  ) async {
+    return GlobalLeftScroll.instance.removeRowAsync(
+      tag,
+      key,
+      onRemove: onRemove,
+    );
+  }
+}
+
+// 为了更短的调用
+class LSTag extends LeftScrollCloseTag {
+  LSTag(String tag) : super(tag);
 }
 
 class LeftScrollCloseTag {
   final String tag;
   const LeftScrollCloseTag(this.tag);
 
+  LeftScrollController of(Key key) => LeftScrollController(
+        tag: this,
+        key: key,
+      );
+
   @override
   int get hashCode => tag.hashCode;
 
-  operator ==(dynamic other) {
+  operator ==(Object other) {
     if (other is LeftScrollCloseTag) {
       if (other.tag == tag) {
         return true;
@@ -82,7 +151,7 @@ class LeftScrollCloseTag {
 enum LeftScrollStatus {
   close,
   open,
-  remove,
+  removing,
   removed,
 }
 
@@ -91,5 +160,5 @@ class LeftScrollStatusCtrl extends ValueNotifier<LeftScrollStatus> {
   LeftScrollStatusCtrl() : super(LeftScrollStatus.close);
   bool get isClose => value == LeftScrollStatus.close;
   bool get isOpen => value == LeftScrollStatus.open;
-  bool get isRemove => value == LeftScrollStatus.remove;
+  bool get isRemove => value == LeftScrollStatus.removing;
 }